Address some review feedback
This commit will move getChplEnv into its own header and source file,
make the Context constructor accept CHPL_HOME and do away with
setChplHome, and make chplHome in Context a std::string.
