+++ compiler/dyno/test/resolution/testInteractive.cpp
@@ -224,7 +224,7 @@ static void usage(int argc, char** argv) {
 }
 
 static void setupSearchPaths(Context* ctx, bool enableStdLib,
-                             const char* chpl_home,
+                             const std::string& chpl_home,
                              const std::vector<std::string>& cmdLinePaths,
                              const std::vector<std::string>& files) {
   if (enableStdLib) {
@@ -255,13 +255,10 @@ static void setupSearchPaths(Context* ctx, bool enableStdLib,
 }
 
 int main(int argc, char** argv) {
-
   bool gc = false;
-  Context context;
-  Context* ctx = &context;
   bool trace = false;
   bool scopeResolveOnly = false;
-  const char* chpl_home = nullptr;
+  std::string chpl_home;
   std::vector<std::string> cmdLinePaths;
   std::vector<std::string> files;
   bool enableStdLib = false;
@@ -287,13 +284,15 @@ int main(int argc, char** argv) {
   if (enableStdLib) {
     if (const char* chpl_home_env  = getenv("CHPL_HOME")) {
       chpl_home = chpl_home_env;
-      context.setChplHome(chpl_home);
       printf("CHPL_HOME is set, so setting up search paths\n");
     } else {
       printf("--std only works when CHPL_HOME is set\n");
       exit(1);
     }
   }
+
+  Context context(chpl_home);
+  Context* ctx = &context;
 
   if (files.size() == 0) {
     usage(argc, argv);
